+++ src/main/java/com/example/todoapp/controller/auth/AuthController.java
@@ -1,11 +1,11 @@
-package com.example.todoapp.controller;
+package com.example.todoapp.controller.auth;
 
 import com.example.todoapp.model.User;
 import com.example.todoapp.payload.JwtAuthResponse;
 import com.example.todoapp.payload.LoginRequest;
 import com.example.todoapp.payload.SignupRequest;
 import com.example.todoapp.security.JwtTokenProvider;
-import com.example.todoapp.service.UserService;
+import com.example.todoapp.service.user.UserService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,9 +15,6 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.HashMap;
-import java.util.Map;
 
 @RestController
 @RequestMapping("/api/auth")
