+++ src/main/java/com/example/todoapp/model/User.java
@@ -21,16 +21,26 @@ public class User {
     @Id
     private String id;
 
+    @NotBlank
+    @Size(min = 3, max = 20)
     @Indexed(unique = true)
     private String username;
 
+    @NotBlank
+    @Size(max = 50)
+    @Email
     @Indexed(unique = true)
     private String email;
 
+    @NotBlank
+    @Size(min = 6, max = 120)
     private String password;
 
     private Set<String> roles = new HashSet<>();
 
-    public User(@NotBlank @Size(min = 3, max = 20) String username, @NotBlank @Size(max = 50) @Email String email, String encode) {
+    public User(String username, String email, String password) {
+        this.username = username;
+        this.email = email;
+        this.password = password;
     }
 }
