+++ src/main/java/com/example/todoapp/service/task/TaskService.java
@@ -1,4 +1,4 @@
-package com.example.todoapp.service;
+package com.example.todoapp.service.task;
 
 import com.example.todoapp.model.Task;
 import org.springframework.data.domain.Page;
@@ -8,17 +8,10 @@ import java.util.List;
 import java.util.Optional;
 
 public interface TaskService {
-    // Admin methods (unchanged)
-    List<Task> getAllTasks();
-    Optional<Task> getTaskById(String id);
-
-    // New user-specific methods
+    // User-specific methods
     List<Task> getUserTasks(String userId);
     Optional<Task> getUserTaskById(String id, String userId);
-
-    // Updated methods to include user ID
     Task createTask(Task task, String userId);
-    Task updateTask(String id, Task task);
     Task updateTask(String id, Task task, String userId);
 
     // User-specific task filtering
@@ -27,17 +20,5 @@ public interface TaskService {
     Page<Task> getPaginatedUserTasks(String userId, Pageable pageable);
     List<Task> searchUserTasksFullText(String searchTerm, String userId);
     Page<Task> getPaginatedUserTasksByStatus(String status, String userId, Pageable pageable);
-    void softDeleteTask(String id);
     void softDeleteTask(String id, String userId);
-
-    // Admin-only methods (unchanged)
-    void restoreTask(String id);
-    void permanentDeleteTask(String id);
-
-    // These methods will need to be maintained for backward compatibility and admin use
-    List<Task> getTasksByStatus(String status);
-    List<Task> searchTasksByTitle(String title);
-    Page<Task> getPaginatedTasks(Pageable pageable);
-    List<Task> searchTasksFullText(String searchTerm);
-    Page<Task> getPaginatedTasksByStatus(String status, Pageable pageable);
 }
