+++ src/main/java/com/example/todoapp/service/user/UserService.java
@@ -1,5 +1,4 @@
-// UserService.java
-package com.example.todoapp.service;
+package com.example.todoapp.service.user;
 
 import com.example.todoapp.model.User;
 import com.example.todoapp.payload.SignupRequest;
