+++ src/main/java/com/example/todoapp/service/user/UserServiceImpl.java
@@ -1,5 +1,4 @@
-// UserServiceImpl.java
-package com.example.todoapp.service;
+package com.example.todoapp.service.user;
 
 import com.example.todoapp.model.User;
 import com.example.todoapp.payload.SignupRequest;
